WIP implement regex match of params with Sentry extension settings, update package.json to add testing
